requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
django-allauth==0.44.0
django-bulma==0.8.1
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
filelock==3.0.12
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==0.10.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tagging==0.2.1
toml==0.10.1
urllib3==1.26.2
virtualenv==20.0.30
wrapt==1.12.1
